Added EDF and LLF

+++ Scheduling/FirstComeFirstServe.cpp
@@ -4,7 +4,7 @@
 
 void FirstComeFirstServe::Schedule()
 {
-	for (size_t i = 0; i < m_Prozesses.size(); i++)
+	for (int i = 0; i < m_Prozesses.size(); i++)
 	{
 		if(m_Prozesses[i].m_timeArrival>m_timePassed)
 		{
